WIP Clean up `aiohttp.AiohttpNotifyServer`/`event_handler.UpnpEventHandler`
