+++ .pre-commit-config.yaml
@@ -55,4 +55,5 @@ repos:
           - defusedxml>=0.6.0
           - pytest==6.2.4
           - voluptuous>=0.12.1
+          - python-didl-lite~=1.3.1
         files: ^(async_upnp_client|tests)/.+\.py$
