+++ async_upnp_client/__init__.py
@@ -9,6 +9,7 @@ from async_upnp_client.client import UpnpService  # noqa: F401
 from async_upnp_client.client import UpnpStateVariable  # noqa: F401
 from async_upnp_client.client_factory import UpnpFactory  # noqa: F401
 from async_upnp_client.event_handler import UpnpEventHandler  # noqa: F401
+from async_upnp_client.event_handler import UpnpNotifyServer  # noqa: F401
 from async_upnp_client.exceptions import UpnpError  # noqa: F401
 from async_upnp_client.exceptions import UpnpValueError  # noqa: F401
 from async_upnp_client.search import SsdpSearchListener  # noqa: F401
