+++ async_upnp_client/event_handler.py
@@ -4,9 +4,9 @@
 import asyncio
 import logging
 import weakref
+from abc import ABC
 from datetime import timedelta
 from http import HTTPStatus
-from socket import AddressFamily  # pylint: disable=no-name-in-module
 from typing import Dict, Mapping, Optional, Tuple, Union
 from urllib.parse import urlparse
 
@@ -20,9 +20,19 @@ from async_upnp_client.exceptions import (
     UpnpResponseError,
     UpnpSIDError,
 )
-from async_upnp_client.utils import async_get_local_ip, get_local_ip
 
 _LOGGER = logging.getLogger(__name__)
+
+
+class UpnpNotifyServer(ABC):
+    """Base Notify Server."""
+
+    # pylint: disable=too-few-public-methods
+
+    @property
+    def callback_url(self) -> str:
+        """Return callback URL on which we are callable."""
+        raise NotImplementedError()
 
 
 class UpnpEventHandler:
@@ -35,24 +45,16 @@ class UpnpEventHandler:
 
     def __init__(
         self,
-        callback_url: str,
+        notify_server: UpnpNotifyServer,
         requester: UpnpRequester,
-        listen_ports: Optional[Mapping[AddressFamily, int]] = None,
     ) -> None:
         """Initialize.
 
-        callback_url can be a normal URL or a format string with {host} and
-            {port} placeholders that will be filled based on how the device is
-            connected.
-        listen_ports is a mapping of IP version to the local listening port,
-            used to determine callback_url for devices.
+        notify_server is the notify server which is actually listening on a socket.
         """
-        self._callback_url = callback_url
+        self._notify_server = notify_server
         self._requester = requester
 
-        self.listen_ports = listen_ports or {}
-
-        self._listen_ip: Optional[str] = None
         self._subscriptions: weakref.WeakValueDictionary[
             str, UpnpService
         ] = weakref.WeakValueDictionary()
@@ -66,34 +68,7 @@ class UpnpEventHandler:
         a device will connect. Use callback_url_for_service to get a more
         specific URL.
         """
-        if not self._listen_ip:
-            self._listen_ip = get_local_ip()
-        port = self.listen_ports.get(AddressFamily.AF_INET)
-        if not port and "{port}" in self._callback_url:
-            raise ValueError("callback_url format requires a listening port")
-        return self._callback_url.format(host=self._listen_ip, port=port)
-
-    async def async_callback_url_for_service(self, service: UpnpService) -> str:
-        """Determine a URL for the service to call back on.
-
-        This can vary based on the service device's IP address.
-        """
-        _LOGGER.debug("Determine callback URL for: %s", service)
-
-        # Shortcut when callback_url can be determined without connecting to the
-        # service device
-        if "{host}" not in self._callback_url and (
-            len(self.listen_ports) == 1 or "{port}" not in self._callback_url
-        ):
-            return self.callback_url
-
-        # Figure out how this host connects to the device, then infer how the
-        # device can connect back
-        device_host = urlparse(service.device.device_url).netloc
-        addr_family, local_host = await async_get_local_ip(device_host)
-        port = self.listen_ports[addr_family]
-
-        return self._callback_url.format(host=local_host, port=port)
+        return self._notify_server.callback_url
 
     def sid_for_service(self, service: UpnpService) -> Optional[str]:
         """Get the service connected to SID."""
@@ -192,16 +167,16 @@ class UpnpEventHandler:
         :raise UpnpCommunicationError (or subclass): Error while performing
             subscription request.
         """
-        callback_url = await self.async_callback_url_for_service(service)
-
-        _LOGGER.debug("Subscribing to: %s, callback URL: %s", service, callback_url)
+        _LOGGER.debug(
+            "Subscribing to: %s, callback URL: %s", service, self.callback_url
+        )
 
         # do SUBSCRIBE request
         headers = {
             "NT": "upnp:event",
             "TIMEOUT": "Second-" + str(timeout.seconds),
             "HOST": urlparse(service.event_sub_url).netloc,
-            "CALLBACK": f"<{callback_url}>",
+            "CALLBACK": f"<{self.callback_url}>",
         }
         response_status, response_headers, _ = await self._requester.async_http_request(
             "SUBSCRIBE", service.event_sub_url, headers
